backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pytesseract
pdf2image
Pillow
requestspydantic>=2.0
google-generativeai